Move `GpuBloomFilterAggregate.scala` into `org.apache.spark.sql.rapids.aggregate`
